Restructure webpack profiles to allow specific hooks
We needed a way to hook right into specific parts of the webpack config,
such as the typescript loader list and the app entry point. It would be
difficult (if not impossible) to do that using only profile merging.
