+++ webpack/test.profile.js
@@ -1,0 +1,25 @@
+const { DefinePlugin } = require('webpack');
+
+module.exports = resolve => ({
+    appEntry: resolve('src/index.dynamic'),
+
+    typescriptLoaders: [
+        {
+            loader: 'awesome-typescript-loader',
+            options: {
+                useBabel: true,
+                useCache: false
+            }
+        },
+        { loader: 'ng-router-loader' }
+    ],
+
+    extraConfig: {
+        devtool:  '#cheap-module-source-map',
+        plugins: [
+            new DefinePlugin({
+                __PRODUCTION__: false
+            })
+        ]
+    }
+});
